refactor: rename async feature to tokio to make it clearer ci: fix cargo check to ensure all targets, features and crates are tested, including examples ci: typos
