+++ .github/workflows/ci.yml
@@ -27,7 +27,7 @@ jobs:
       - uses: dtolnay/rust-toolchain@stable
       - uses: Swatinem/rust-cache@v2
       - name: Run cargo test
-        run: cargo test --all
+        run: cargo test --all --all-targets --all-features
       - name: Run cargo doc
         run: cargo doc --no-deps
 
@@ -45,4 +45,4 @@ jobs:
           toolchain: ${{ steps.msrv.outputs.version }}
       - uses: Swatinem/rust-cache@v2
       - name: Run cargo check
-        run: cargo check
+        run: cargo check --all-targets --all-features
