+++ .pre-commit-config.yaml
@@ -13,6 +13,10 @@ repos:
       - id: check-symlinks
       - id: mixed-line-ending
         args: ["--fix=lf"]
+  - repo: https://github.com/crate-ci/typos
+    rev: v1.18.0
+    hooks:
+      - id: typos
   - repo: https://github.com/doublify/pre-commit-rust
     rev: v1.0
     hooks:
