+++ _typos.toml
@@ -1,0 +1,100 @@
+[default]
+
+[default.extend-words]
+# Don't correct LFS Packet names
+Isi = "Isi"
+Ver = "Ver"
+Tiny = "Tiny"
+Small = "Small"
+Sta = "Sta"
+Sch = "Sch"
+Sfp = "Sfp"
+Scc = "Scc"
+Cpp = "Cpp"
+Ism = "Ism"
+Mso = "Mso"
+Iii = "Iii"
+Mst = "Mst"
+Mtc = "Mtc"
+Mod = "Mod"
+Vtn = "Vtn"
+Rst = "Rst"
+Ncn = "Ncn"
+Cnl = "Cnl"
+Cpr = "Cpr"
+Npl = "Npl"
+Plp = "Plp"
+Pll = "Pll"
+Lap = "Lap"
+Spx = "Spx"
+Pit = "Pit"
+Psf = "Psf"
+Pla = "Pla"
+Cch = "Cch"
+Pen = "Pen"
+Toc = "Toc"
+Flg = "Flg"
+Pfl = "Pfl"
+Fin = "Fin"
+Res = "Res"
+Reo = "Reo"
+Nlp = "Nlp"
+Mci = "Mci"
+Msx = "Msx"
+Msl = "Msl"
+Crs = "Crs"
+Bfn = "Bfn"
+Axi = "Axi"
+Axo = "Axo"
+Btn = "Btn"
+Btc = "Btc"
+Btt = "Btt"
+Rip = "Rip"
+Ssh = "Ssh"
+Con = "Con"
+Obh = "Obh"
+Hlv = "Hlv"
+Plc = "Plc"
+Axm = "Axm"
+Acr = "Acr"
+Hcp = "Hcp"
+Nci = "Nci"
+Jrr = "Jrr"
+Uco = "Uco"
+Oco = "Oco"
+Ttc = "Ttc"
+Slc = "Slc"
+Csc = "Csc"
+Cim = "Cim"
+Mal = "Mal"
+Plh = "Plh"
+RelayArq = "RelayArq"
+RelayArp = "RelayArp"
+RelayHlr = "RelayHlr"
+RelayHos = "RelayHos"
+RelaySel = "RelaySel"
+RelayErr = "RelayErr"
+
+# Don't correct LFS Vehicles
+Xfg = "Xfg"
+Xrg = "Xrg"
+Fbm = "Fbm"
+Xrt = "Xrt"
+Rb4 = "Rb4"
+Fxo = "Fxo"
+Lx4 = "Lx4"
+Lx6 = "Lx6"
+Mrt = "Mrt"
+Uf1 = "Uf1"
+Rac = "Rac"
+Fz5 = "Fz5"
+Fox = "Fox"
+Xfr = "Xfr"
+Ufr = "Ufr"
+Fo8 = "Fo8"
+Fxr = "Fxr"
+Xrr = "Xrr"
+Fzr = "Fzr"
+Bf1 = "Bf1"
+
+Fo = "Fo"
