+++ insim/src/insim/toc.rs
@@ -7,7 +7,7 @@ use crate::identifiers::{ConnectionId, PlayerId, RequestId};
 #[cfg_attr(feature = "serde", derive(serde::Serialize))]
 /// Take Over Car - informational - when a 2 connections swap drivers
 /// Insim indicates this by sending this packet which describes a transfer of the relationship
-/// betwen this PlayerId and two ConnectionId's.
+/// between this PlayerId and two ConnectionId's.
 pub struct Toc {
     /// Non-zero if the packet is a packet request or a reply to a request
     pub reqi: RequestId,
