+++ insim/src/net/blocking_impl/mod.rs
@@ -1,30 +1,88 @@
-use std::{fmt::Debug, time::Duration};
-
-use bytes::BytesMut;
-use if_chain::if_chain;
-use tokio::{
-    io::BufWriter,
+use std::{
+    fmt::Debug,
+    io::{Read, Write},
     net::{TcpStream, UdpSocket},
-    time::{self, timeout},
+    time::Duration,
 };
 
-#[cfg(feature = "websocket")]
-use super::websocket::TungsteniteWebSocket;
-use super::AsyncTryReadWriteBytes;
-use crate::{
-    error::Error,
-    insim::Isi,
-    net::{Codec, DEFAULT_TIMEOUT_SECS},
-    packet::Packet,
-    result::Result,
-};
+use bytes::{BufMut, BytesMut};
 
-/// A unified wrapper around anything that implements [TryReadWriteBytes].
+use super::{Codec, DEFAULT_TIMEOUT_SECS};
+use crate::{insim::Isi, result::Result, Error, Packet, DEFAULT_BUFFER_CAPACITY};
+
+/// TryReadWriteBytes
+pub trait TryReadWriteBytes {
+    /// Set the timeout
+    fn try_read_write_timeout(&self, timeout: Duration);
+
+    /// Try to read bytes.
+    fn try_read_bytes(&mut self, buf: &mut BytesMut) -> Result<usize>;
+
+    /// Try to write bytes.
+    fn try_write_bytes(&mut self, src: &[u8]) -> Result<usize>;
+}
+
+impl TryReadWriteBytes for TcpStream {
+    fn try_read_write_timeout(&self, timeout: Duration) {
+        self.set_read_timeout(Some(timeout))
+            .expect("set_read_timeout failed");
+        self.set_write_timeout(Some(timeout))
+            .expect("set_write_timeout failed");
+    }
+
+    fn try_read_bytes(&mut self, buf: &mut BytesMut) -> Result<usize> {
+        // TODO: Remove when read_buf becomes stable.
+        // See https://users.rust-lang.org/t/how-to-read-from-tcpstream-and-append-to-vec-u8-efficiently/89059/4
+        // for why this is "safe"
+
+        let size = {
+            let chunk = buf.chunk_mut();
+            let len = chunk.len();
+
+            let ptr = chunk.as_mut_ptr();
+            // SAFETY: The file is not going to read from the uninitialized data.
+            let slice = unsafe { std::slice::from_raw_parts_mut(ptr, len) };
+            self.read(slice)?
+        };
+
+        // SAFETY: The tcpstream just initialized that many bytes.
+        unsafe {
+            buf.advance_mut(size);
+        }
+        Ok(size)
+    }
+
+    fn try_write_bytes(&mut self, src: &[u8]) -> Result<usize> {
+        Ok(self.write(src)?)
+    }
+}
+
+impl TryReadWriteBytes for UdpSocket {
+    fn try_read_write_timeout(&self, timeout: Duration) {
+        self.set_read_timeout(Some(timeout))
+            .expect("set_read_timeout failed");
+        self.set_write_timeout(Some(timeout))
+            .expect("set_write_timeout failed");
+    }
+
+    fn try_read_bytes(&mut self, buf: &mut BytesMut) -> Result<usize> {
+        let mut rx_bytes = [0u8; crate::MAX_SIZE_PACKET];
+        let size = self.recv(&mut rx_bytes)?;
+        buf.extend_from_slice(&rx_bytes[..size]);
+        Ok(size)
+    }
+
+    fn try_write_bytes(&mut self, src: &[u8]) -> Result<usize> {
+        Ok(self.send(src)?)
+    }
+}
+
+/// A unified wrapper around anything that implements Read + Write.
 /// You probably really want to look at [Framed].
 #[derive(Debug)]
 pub struct FramedInner<N>
 where
-    N: AsyncTryReadWriteBytes,
+    N: TryReadWriteBytes,
 {
     inner: N,
     codec: Codec,
@@ -34,11 +92,12 @@ where
 
 impl<N> FramedInner<N>
 where
-    N: AsyncTryReadWriteBytes,
+    N: TryReadWriteBytes,
 {
     /// Create a new FramedInner, which wraps some kind of network transport.
     pub fn new(inner: N, codec: Codec) -> Self {
-        let buffer = BytesMut::new();
+        let buffer = BytesMut::with_capacity(DEFAULT_BUFFER_CAPACITY);
+        inner.try_read_write_timeout(Duration::from_secs(DEFAULT_TIMEOUT_SECS));
 
         Self {
             inner,
@@ -56,16 +115,16 @@ where
     /// Performs the Insim handshake by sending a [Isi] packet.
     /// If the handshake does not complete within the given timeout, it will fail and the
     /// connection should be considered invalid.
-    pub async fn handshake(&mut self, isi: Isi, timeout: Duration) -> Result<()> {
-        time::timeout(timeout, self.write(isi.into())).await??;
+    pub fn handshake(&mut self, isi: Isi) -> Result<()> {
+        self.write(isi.into())?;
 
         Ok(())
     }
 
-    /// Asynchronously wait for a packet from the inner network.
-    pub async fn read(&mut self) -> Result<Packet> {
+    /// Wait for a packet from the inner network.
+    pub fn read(&mut self) -> Result<Packet> {
         loop {
-            if_chain! {
+            if_chain::if_chain! {
                 if !self.buffer.is_empty();
                 if let Some(packet) = self.codec.decode(&mut self.buffer)?;
                 then {
@@ -77,19 +136,14 @@ where
                     // keepalive
                     if let Some(pong) = packet.maybe_pong() {
                         tracing::debug!("Ping? Pong!");
-                        self.write(pong).await?;
+                        self.write(pong)?;
                     }
 
                     return Ok(packet);
                 }
             }
 
-            match timeout(
-                Duration::from_secs(DEFAULT_TIMEOUT_SECS),
-                self.inner.try_read_bytes(&mut self.buffer),
-            )
-            .await?
-            {
+            match self.inner.try_read_bytes(&mut self.buffer) {
                 Ok(0) => {
                     // The remote closed the connection. For this to be a clean
                     // shutdown, there should be no data in the read buffer. If
@@ -108,17 +162,18 @@ where
                     continue;
                 },
                 Err(e) => {
+                    tracing::info!("did get err={:?}", e);
                     return Err(e);
                 },
             }
         }
     }
 
-    /// Asynchronously write a packet to the inner network.
-    pub async fn write(&mut self, packet: Packet) -> Result<()> {
+    /// Write a packet to the inner network.
+    pub fn write(&mut self, packet: Packet) -> Result<()> {
         let buf = self.codec.encode(&packet)?;
         if !buf.is_empty() {
-            let _ = self.inner.try_write_bytes(&buf).await?;
+            let _ = self.inner.try_write_bytes(&buf)?;
         }
 
         Ok(())
@@ -134,40 +189,30 @@ where
 pub enum Framed {
     /// Tcp
     Tcp(FramedInner<TcpStream>),
-    /// BufferedTcp
-    BufferedTcp(FramedInner<BufWriter<TcpStream>>),
+
     /// Udp
     Udp(FramedInner<UdpSocket>),
-    #[cfg(feature = "websocket")]
-    /// Websocket, primarily intended for use with the LFS World relay.
-    WebSocket(FramedInner<TungsteniteWebSocket>),
 }
 
 impl Framed {
     #[tracing::instrument]
-    /// Asynchronously wait for a packet from the inner network.
-    pub async fn read(&mut self) -> Result<Packet> {
+    /// Wait for a packet from the inner network.
+    pub fn read(&mut self) -> Result<Packet> {
         let res = match self {
-            Self::Tcp(i) => i.read().await,
-            Self::Udp(i) => i.read().await,
-            Self::BufferedTcp(i) => i.read().await,
-            #[cfg(feature = "websocket")]
-            Self::WebSocket(i) => i.read().await,
+            Self::Tcp(i) => i.read(),
+            Self::Udp(i) => i.read(),
         };
         tracing::debug!("read result {:?}", res);
         res
     }
 
     #[tracing::instrument]
-    /// Asynchronously write a packet to the inner network.
-    pub async fn write<I: Into<Packet> + Send + Sync + Debug>(&mut self, packet: I) -> Result<()> {
+    /// Write a packet to the inner network.
+    pub fn write<I: Into<Packet> + Send + Sync + Debug>(&mut self, packet: I) -> Result<()> {
         tracing::debug!("writing packet {:?}", &packet);
         match self {
-            Self::Tcp(i) => i.write(packet.into()).await,
-            Self::Udp(i) => i.write(packet.into()).await,
-            Self::BufferedTcp(i) => i.write(packet.into()).await,
-            #[cfg(feature = "websocket")]
-            Self::WebSocket(i) => i.write(packet.into()).await,
+            Self::Tcp(i) => i.write(packet.into()),
+            Self::Udp(i) => i.write(packet.into()),
         }
     }
 }
@@ -180,19 +225,7 @@ impl Debug for Framed {
                 "Framed::Tcp {{ codec: {:?}, verify_version: {:?} }}",
                 i.codec, i.verify_version
             ),
-            Framed::BufferedTcp(i) => write!(
-                f,
-                "Framed::BufferedTcp {{ codec: {:?}, verify_version: {:?} }}",
-                i.codec, i.verify_version
-            ),
             Framed::Udp(i) => write!(
-                f,
-                "Framed::Tcp {{ codec: {:?}, verify_version: {:?} }}",
-                i.codec, i.verify_version
-            ),
-
-            #[cfg(feature = "websocket")]
-            Framed::WebSocket(i) => write!(
                 f,
                 "Framed::Tcp {{ codec: {:?}, verify_version: {:?} }}",
                 i.codec, i.verify_version
