+++ insim/src/relay/admin_response.rs
@@ -2,7 +2,7 @@ use insim_core::binrw::{self, binrw};
 
 use crate::identifiers::RequestId;
 
-/// Reponse to a [super::admin_request::Arq] packet, indicating if we are logged in as an administrative user on
+/// Response to a [super::admin_request::Arq] packet, indicating if we are logged in as an administrative user on
 /// the selected host.
 #[binrw]
 #[derive(Debug, Clone, Default)]
